requirements.txt
torch
transformers
openai
evaluate
tqdm
pandas
rouge-score
PyYAML